Initial split team and solo
