+++ src/AutoMerge/RecentChangesets/Solo/RecentChangesetsSoloView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="AutoMerge.RecentChangesetsView"
+<UserControl x:Class="AutoMerge.RecentChangesetsSoloView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:autoMerge="clr-namespace:AutoMerge;assembly="
