+++ src/AutoMerge/RecentChangesets/Team/RecentChangesetsTeamView.xaml
@@ -1,0 +1,82 @@
+<UserControl x:Class="AutoMerge.RecentChangesetsTeamView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:autoMerge="clr-namespace:AutoMerge;assembly="
+             xmlns:converters="clr-namespace:Microsoft.TeamFoundation.Controls.WPF.Converters;assembly=Microsoft.TeamFoundation.Controls"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:i="clr-namespace:System.Windows.Interactivity;assembly=System.Windows.Interactivity"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:teamExplorer="clr-namespace:Microsoft.TeamFoundation.Controls.WPF.TeamExplorer;assembly=Microsoft.TeamFoundation.Controls"
+             xmlns:wpf="clr-namespace:Microsoft.TeamFoundation.Controls.WPF;assembly=Microsoft.TeamFoundation.Controls"
+             TextOptions.TextFormattingMode="Display"
+             UseLayoutRounding="True"
+             d:DataContext="{d:DesignInstance autoMerge:RecentChangesetsTeamViewModel}"
+             d:DesignHeight="300"
+             d:DesignWidth="300"
+             mc:Ignorable="d">
+  <UserControl.Resources>
+    <ResourceDictionary>
+      <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="pack://application:,,,/Microsoft.TeamFoundation.Controls;component/WPF/SharedResources.xaml" />
+      </ResourceDictionary.MergedDictionaries>
+
+      <autoMerge:ChangesetCommentConverter x:Key="ChangesetCommentConverter" />
+      <converters:NonEmptyStringToVisibleConverter x:Key="NonEmptyStringToVisibleConverter" />
+      <DataTemplate x:Key="ChangesetItemDataTemplate" DataType="autoMerge:ChangesetViewModel">
+        <StackPanel Orientation="Horizontal">
+          <StackPanel.ToolTip>
+            <ToolTip Content="{Binding Comment}" />
+          </StackPanel.ToolTip>
+          <TextBlock Margin="0, 0, 0, 0" Text="{Binding ChangesetId}" />
+          <TextBlock Margin="5, 0, 0, 0"
+                     VerticalAlignment="Center"
+                     Background="{DynamicResource EmbeddedDialogBackgroundBrushKey}"
+                     FontSize="8"
+                     Foreground="{DynamicResource BodyTextBrushKey}"
+                     Padding="2"
+                     Text="{Binding DisplayBranchName}"
+                     Visibility="{Binding DisplayBranchName,
+                                          Converter={StaticResource NonEmptyStringToVisibleConverter}}" />
+          <TextBlock Margin="10, 0, 0, 0" Text="{Binding Comment, Converter={StaticResource ChangesetCommentConverter}}" />
+        </StackPanel>
+      </DataTemplate>
+      <Style x:Key="ChangesetListStyle"
+             BasedOn="{StaticResource {x:Type ListBox}}"
+             TargetType="ListBox">
+        <Setter Property="ItemTemplate" Value="{StaticResource ChangesetItemDataTemplate}" />
+        <Setter Property="ItemContainerStyle" Value="{StaticResource ListBoxItemStyle}" />
+        <Setter Property="SelectionMode" Value="Single" />
+        <Setter Property="ScrollViewer.HorizontalScrollBarVisibility" Value="Hidden" />
+        <Setter Property="ScrollViewer.VerticalScrollBarVisibility" Value="Auto" />
+        <Setter Property="BorderThickness" Value="0" />
+        <Setter Property="IsSynchronizedWithCurrentItem" Value="True" />
+      </Style>
+    </ResourceDictionary>
+  </UserControl.Resources>
+
+  <StackPanel Margin="6,6,6,6">
+    <ListBox x:Name="changesetList"
+             Margin="0"
+             ItemsSource="{Binding Changesets}"
+             SelectedItem="{Binding SelectedChangeset,
+                                    Mode=TwoWay}"
+             Style="{StaticResource ChangesetListStyle}"
+             SelectionMode="Multiple">
+      <ListBox.InputBindings>
+        <KeyBinding Key="Enter" Command="{Binding ViewChangesetDetailsCommand}" />
+      </ListBox.InputBindings>
+      <i:Interaction.Triggers>
+        <i:EventTrigger EventName="MouseDoubleClick">
+          <i:InvokeCommandAction Command="{Binding ViewChangesetDetailsCommand}" />
+        </i:EventTrigger>
+      </i:Interaction.Triggers>
+      <!--
+        Comment adding behavior, because in some cases error
+        ' Cannot add instance of type 'ScrollDeligateBehavior' to a collection of type 'BehaviorCollection'. Only items of type 'T' are allowed.'
+      -->
+      <!--  <i:Interaction.Behaviors>  -->
+      <!--  <behaviours:ScrollDeligateBehavior />  -->
+      <!--  </i:Interaction.Behaviors>  -->
+    </ListBox>
+  </StackPanel>
+</UserControl>
